Bugfix/incorrect count on search books

+++ src/main/java/kr/milibrary/domain/BookList.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class BookList extends BaseDomain {
-    private Integer totalPage = null;
+    private Long totalPage = null;
     private List<Book> bookList;
 
     public BookList() {
@@ -17,16 +17,16 @@ public class BookList extends BaseDomain {
         this.bookList = bookList;
     }
 
-    public BookList(Integer totalPage, List<Book> bookList) {
+    public BookList(Long totalPage, List<Book> bookList) {
         this.totalPage = totalPage;
         this.bookList = bookList;
     }
 
-    public Integer getTotalPage() {
+    public Long getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Integer totalPage) {
+    public void setTotalPage(Long totalPage) {
         this.totalPage = totalPage;
     }
